[SQL] Count records that refer to a specific object
eoli_addons.count_row_referrers (src_tab TEXT, rec_id INT)
eoli_addons.count_row_referrers (src_tab TEXT, rec_id TEXT)

Counts the number of rows that refer to a specified record such as a
user or an organizational unit. Two versions of the function exist,
one that assumes that the record ID is an integer and one that
assumes it is a string.

These functions (currently) only include tables that refer to the
source object via direct foreign key relationships. This means that
any relationships that refer to the source object indirectly
(e.g., in a library setting or an item template) are not counted.

Example to identify tables that refer to organizational unit 4:

SELECT * FROM eoli_addons.count_row_referrers('actor.org_unit', 4) WHERE num_rows > 0;

Arguments:

  1. The schema-qualfied name of the table that has the source object
  2. The record ID of the source object

Returns: list of the tables that refer to that object, including column and record count

Signed-off-by: Galen Charlton <gmc@equinoxOLI.org>